# In PostgreSQL, allow spaces in table names without schema