--- a/activerecord/test/cases/adapters/postgresql/utils_test.rb
+++ b/activerecord/test/cases/adapters/postgresql/utils_test.rb
@@ -14,6 +14,8 @@ class PostgreSQLUtilsTest < ActiveRecord::PostgreSQLTestCase
       %(schema."table_name")   => %w{schema table_name},
       %("schema"."table_name") => %w{schema table_name},
       %("even spaces".table)   => ["even spaces","table"],
+      %(spaces table only)     => [nil,"spaces table only"],
+      %("spaces table only")   => [nil,"spaces table only"],
       %(schema."table.name")   => ["schema", "table.name"]
     }.each do |given, expect|
       assert_equal Name.new(*expect), extract_schema_qualified_name(given)