@matthewd 2016-10-06
I don't see why we'd want to treat `"x.foo bar"` differently from `"foo bar"`.

The current intention appears to be that the table name is treated as an unquoted identifier, meaning `"\"foo bar\""` would already work correctly.