@Jerph 2016-10-06
If so, it's inconsistent with the mysql/sqlite adapter behavior, which appear to handle table names with spaces as a quoted identifier and have no problem with just "foo bar". Then again, they are the outliers in not having separate schema/database concepts, so maybe there's an argument for different behavior.

Honestly, I don't understand why the whitespace is checked in that regex at all. Is it intentionally another option for specifying schema? White space isn't shown in any of the test cases, and the only way I know to specify schema is with a dot. 

This fix doesn't preclude handling the parsing logic any way we want, so actual table names with dots in them (lord help us :-D) are still handled correctly. If there's no dot though, there simply can't be a schema, so it should be treated as a quoted identifier, in my opinion.