@matthewd 2016-10-06
Looks like it's been an undocumented behaviour since #1632.

As is, this would change (read: break) handling of `"\"foo\""`, in an IMO undesirable way.

Let's go the safest route, and deprecate unquoted spaces -- leave it parsing as a schema for now (and users can quote for the table-name-with-spaces interpretation), then look at stricter options later.