@Jerph 2016-10-06
Good point! I added a test for quoted strings without a schema, and I don't believe this changes any behavior with "\"foo\"" because strings are unquoted by the Name class.

The different behavior should only be with quotes and spaces when there's no dot in the string. Malformed double quotes for instance. Some examples I can think of:

```
"foo bar" -> [nil, "foo bar"] vs ["foo", "bar"] # the original problem
"foo\"bar" -> [nil, "foo\"bar"] vs ["foo", "bar"] # unquote ignores quotes in the middle
"\"foo\" bar" -> [nil, "foo\" ba"] vs ["foo", "bar"] # unquote strips off the last character, assuming a paired quote
```
All of these examples were erroneous before, though. I definitely won't say that there are no legitimate identifiers that would be broken, but I can't think of any.

Regarding #1632, I didn't see any reasoning for the use of `\s` in the regex. If it could be removed that possibly be a better solution. As it is, I can't imagine any code running successfully that relies on the split on white space.