@matthewd 2016-10-06
> `"foo bar" -> [nil, "foo bar"] vs ["foo", "bar"] # the original problem`

This doesn't seem sufficiently far-fetched to me that we can silently change it. (The others do seem inconsequential.)

So, I agree the `\s` should go away, but do believe we need to give warning.