@Jerph 2016-10-06
That sounds good. Thanks for helping me on this.